requirements.txt
django
django-cleanup
django-cors-headers
djangorestframework
gunicorn
numpy
Pillow
scikit-learn
scipy
pandas
requests